requirements.txt
openai>=1.0.0
python-dotenv==1.2.2
spotipy==2.26.0
tqdm==4.67.3
